Safely get currentUserState
