+++ frontend/src/pages/notebookController/SetupCurrentUserState.tsx
@@ -1,0 +1,56 @@
+import * as React from 'react';
+import { NotebookControllerContext } from './NotebookControllerContext';
+import ApplicationsPage from '../ApplicationsPage';
+import {
+  getUserStateFromDashboardConfig,
+  usernameTranslate,
+} from '../../utilities/notebookControllerUtils';
+import { patchDashboardConfig } from '../../services/dashboardConfigService';
+import AppContext from '../../app/AppContext';
+import { useUser } from '../../redux/selectors';
+import { EMPTY_USER_STATE } from './const';
+
+const SetupCurrentUserState: React.FC = ({ children }) => {
+  const { dashboardConfig } = React.useContext(AppContext);
+  const { currentUserState, setCurrentUserState } = React.useContext(NotebookControllerContext);
+  const { username } = useUser();
+  const translatedUsername = usernameTranslate(username);
+
+  React.useEffect(() => {
+    if (!currentUserState.user) {
+      // Load the user's current state for the NotebookController
+      if (translatedUsername && dashboardConfig.spec.notebookController) {
+        const notebookControllerState = dashboardConfig.status?.notebookControllerState || [];
+        let fetchedUserState = getUserStateFromDashboardConfig(
+          translatedUsername,
+          notebookControllerState,
+        );
+
+        if (!fetchedUserState) {
+          fetchedUserState = {
+            ...EMPTY_USER_STATE,
+            user: username,
+          };
+          const patch = {
+            status: {
+              notebookControllerState: notebookControllerState
+                ? [...notebookControllerState, currentUserState]
+                : [currentUserState],
+            },
+          };
+          patchDashboardConfig(patch).catch((e) => console.error(e));
+        }
+
+        setCurrentUserState(fetchedUserState);
+      }
+    }
+  }, [translatedUsername, dashboardConfig, currentUserState, setCurrentUserState, username]);
+
+  if (!currentUserState?.user) {
+    return <ApplicationsPage title="Loading..." description={null} loaded={false} empty={false} />;
+  }
+
+  return <>{children}</>;
+};
+
+export default SetupCurrentUserState;
